refactor: move .label() to a LabelUtils file
