+++ composeApp/src/commonMain/kotlin/paige/navic/ui/components/sheets/SongSheet.kt
@@ -73,6 +73,7 @@ import paige.navic.icons.outlined.Star
 import paige.navic.managers.SleepTimerManager
 import paige.navic.ui.components.common.CoverArt
 import paige.navic.ui.theme.positive
+import paige.navic.utils.label
 
 @OptIn(ExperimentalMaterial3Api::class, ExperimentalMaterial3ExpressiveApi::class)
 @Composable
